# Mimir query engine: return an error when a query produces multiple series with the same labels