--- a/pkg/streamingpromql/functions/range_vectors.go
+++ b/pkg/streamingpromql/functions/range_vectors.go
@@ -17,8 +17,9 @@ import (
 )
 
 var CountOverTime = FunctionOverRangeVector{
-	SeriesMetadataFunc: DropSeriesName,
-	StepFunc:           countOverTime,
+	SeriesMetadataFunc:       DropSeriesName,
+	NeedsSeriesDeduplication: true,
+	StepFunc:                 countOverTime,
 }
 
 func countOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPointRingBuffer, hPoints *types.HPointRingBuffer, _ EmitAnnotationFunc) (float64, bool, *histogram.FloatHistogram, error) {
@@ -54,8 +55,9 @@ func lastOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPoi
 }
 
 var PresentOverTime = FunctionOverRangeVector{
-	SeriesMetadataFunc: DropSeriesName,
-	StepFunc:           presentOverTime,
+	SeriesMetadataFunc:       DropSeriesName,
+	NeedsSeriesDeduplication: true,
+	StepFunc:                 presentOverTime,
 }
 
 func presentOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPointRingBuffer, hPoints *types.HPointRingBuffer, _ EmitAnnotationFunc) (float64, bool, *histogram.FloatHistogram, error) {
@@ -67,8 +69,9 @@ func presentOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.F
 }
 
 var MaxOverTime = FunctionOverRangeVector{
-	SeriesMetadataFunc: DropSeriesName,
-	StepFunc:           maxOverTime,
+	SeriesMetadataFunc:       DropSeriesName,
+	NeedsSeriesDeduplication: true,
+	StepFunc:                 maxOverTime,
 }
 
 func maxOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPointRingBuffer, _ *types.HPointRingBuffer, _ EmitAnnotationFunc) (float64, bool, *histogram.FloatHistogram, error) {
@@ -104,8 +107,9 @@ func maxOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPoin
 }
 
 var MinOverTime = FunctionOverRangeVector{
-	SeriesMetadataFunc: DropSeriesName,
-	StepFunc:           minOverTime,
+	SeriesMetadataFunc:       DropSeriesName,
+	NeedsSeriesDeduplication: true,
+	StepFunc:                 minOverTime,
 }
 
 func minOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPointRingBuffer, _ *types.HPointRingBuffer, _ EmitAnnotationFunc) (float64, bool, *histogram.FloatHistogram, error) {
@@ -142,6 +146,7 @@ func minOverTime(step types.RangeVectorStepData, _ float64, fPoints *types.FPoin
 
 var SumOverTime = FunctionOverRangeVector{
 	SeriesMetadataFunc:             DropSeriesName,
+	NeedsSeriesDeduplication:       true,
 	StepFunc:                       sumOverTime,
 	NeedsSeriesNamesForAnnotations: true,
 }
@@ -217,6 +222,7 @@ func sumHistograms(head, tail []promql.HPoint, emitAnnotation EmitAnnotationFunc
 
 var AvgOverTime = FunctionOverRangeVector{
 	SeriesMetadataFunc:             DropSeriesName,
+	NeedsSeriesDeduplication:       true,
 	StepFunc:                       avgOverTime,
 	NeedsSeriesNamesForAnnotations: true,
 }