#### What this PR does

This PR introduces a check in MQE for queries that produce results with multiple series with the same labels.

This is currently only possible in MQE in places where we drop the metric name from the input series: in all other cases it's not possible to return multiple series with the same labels. (Functions like `label_replace` and unary negation will need similar checks when we implement those features.)

Latency is slightly worse in some cases, and slightly better in others, which I'm putting down to benchmark noise rather than anything else:

<details>
<summary>Benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
                                                                                                                  │   main.txt   │              after.txt              │
                                                                                                                  │    sec/op    │    sec/op     vs base               │
Query/rate(a_1[1m]),_instant_query/Mimir-10                                                                         144.4µ ±  8%   164.4µ ±  5%  +13.86% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps/Mimir-10                                                            159.5µ ±  4%   178.1µ ±  3%  +11.62% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps/Mimir-10                                                           218.8µ ±  7%   241.1µ ±  3%  +10.18% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query/Mimir-10                                                                       413.1µ ±  2%   459.2µ ±  3%  +11.17% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps/Mimir-10                                                          1.455m ±  8%   1.566m ±  5%   +7.60% (p=0.041 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps/Mimir-10                                                         6.200m ±  6%   6.606m ±  3%   +6.55% (p=0.026 n=6)
Query/rate(a_2000[1m]),_instant_query/Mimir-10                                                                      4.109m ±  6%   4.270m ±  3%        ~ (p=0.180 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps/Mimir-10                                                         22.45m ±  1%   23.86m ±  5%   +6.25% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps/Mimir-10                                                        111.5m ±  3%   119.2m ±  6%   +6.90% (p=0.009 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps/Mimir-10                                                          820.9µ ±  1%   814.4µ ±  0%   -0.79% (p=0.004 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps/Mimir-10                                                        56.73m ±  1%   57.28m ±  2%   +0.96% (p=0.015 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps/Mimir-10                                                        1.130 ±  7%    1.131 ±  6%        ~ (p=0.699 n=6)
Query/rate(nh_1[1m]),_instant_query/Mimir-10                                                                        137.9µ ±  5%   135.4µ ±  0%        ~ (p=0.240 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/Mimir-10                                                           272.4µ ±  7%   267.7µ ±  0%   -1.72% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/Mimir-10                                                          654.4µ ±  0%   655.1µ ±  2%        ~ (p=1.000 n=6)
Query/rate(nh_100[1m]),_instant_query/Mimir-10                                                                      566.4µ ±  3%   563.7µ ±  1%        ~ (p=0.132 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/Mimir-10                                                         11.17m ±  4%   11.08m ±  0%   -0.82% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/Mimir-10                                                        43.12m ±  1%   42.76m ±  1%        ~ (p=0.093 n=6)
Query/rate(nh_2000[1m]),_instant_query/Mimir-10                                                                     6.368m ±  4%   6.407m ±  1%        ~ (p=0.240 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/Mimir-10                                                        207.2m ±  1%   208.6m ±  2%   +0.67% (p=0.026 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/Mimir-10                                                       834.0m ±  2%   833.1m ±  2%        ~ (p=0.818 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/Mimir-10                                                         4.618m ±  3%   4.551m ±  2%        ~ (p=0.065 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/Mimir-10                                                       409.9m ±  6%   409.3m ±  1%        ~ (p=0.310 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/Mimir-10                                                       8.588 ±  2%    8.589 ±  1%        ~ (p=0.818 n=6)
Query/rate(a_1[1d]),_instant_query/Mimir-10                                                                         586.8µ ±  2%   579.9µ ±  1%        ~ (p=0.180 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps/Mimir-10                                                            936.2µ ±  3%   921.3µ ±  1%   -1.59% (p=0.015 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps/Mimir-10                                                           3.884m ±  1%   3.824m ±  0%   -1.56% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query/Mimir-10                                                                       34.71m ±  3%   34.50m ±  1%        ~ (p=0.180 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps/Mimir-10                                                          67.52m ±  6%   67.56m ±  1%        ~ (p=1.000 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps/Mimir-10                                                         363.4m ±  4%   360.1m ±  1%   -0.89% (p=0.015 n=6)
Query/rate(a_2000[1d]),_instant_query/Mimir-10                                                                      641.2m ±  2%   636.2m ±  3%        ~ (p=0.132 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps/Mimir-10                                                          1.311 ±  3%    1.300 ±  1%   -0.84% (p=0.041 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps/Mimir-10                                                         7.183 ±  0%    7.176 ±  0%        ~ (p=0.589 n=6)
Query/rate(nh_1[1h]),_instant_query/Mimir-10                                                                        277.9µ ±  1%   279.5µ ±  4%        ~ (p=0.180 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/Mimir-10                                                           441.4µ ±  5%   432.3µ ±  2%   -2.05% (p=0.041 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/Mimir-10                                                          1.810m ±  4%   1.759m ±  0%   -2.82% (p=0.002 n=6)
Query/rate(nh_100[1h]),_instant_query/Mimir-10                                                                      8.464m ±  4%   8.356m ±  1%   -1.28% (p=0.004 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/Mimir-10                                                         21.74m ±  2%   21.09m ±  0%   -2.99% (p=0.002 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/Mimir-10                                                        152.4m ±  1%   150.4m ±  0%   -1.36% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_instant_query/Mimir-10                                                                     161.9m ±  1%   159.8m ±  1%   -1.30% (p=0.009 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/Mimir-10                                                        408.8m ±  1%   405.8m ±  1%   -0.73% (p=0.015 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/Mimir-10                                                        3.044 ±  1%    3.020 ±  1%   -0.80% (p=0.041 n=6)
Query/avg_over_time(a_1[1m]),_instant_query/Mimir-10                                                                131.3µ ±  6%   127.8µ ±  6%        ~ (p=0.132 n=6)
Query/avg_over_time(a_1[1m]),_range_query_with_100_steps/Mimir-10                                                   153.7µ ±  3%   145.2µ ±  3%   -5.52% (p=0.009 n=6)
Query/avg_over_time(a_1[1m]),_range_query_with_1000_steps/Mimir-10                                                  223.2µ ±  7%   218.7µ ±  1%   -2.04% (p=0.009 n=6)
Query/avg_over_time(a_100[1m]),_instant_query/Mimir-10                                                              390.7µ ±  3%   388.6µ ±  2%        ~ (p=0.589 n=6)
Query/avg_over_time(a_100[1m]),_range_query_with_100_steps/Mimir-10                                                 1.562m ±  0%   1.627m ±  5%   +4.14% (p=0.002 n=6)
Query/avg_over_time(a_100[1m]),_range_query_with_1000_steps/Mimir-10                                                7.545m ±  2%   7.551m ±  1%        ~ (p=1.000 n=6)
Query/avg_over_time(a_2000[1m]),_instant_query/Mimir-10                                                             4.190m ±  3%   4.241m ±  0%        ~ (p=0.065 n=6)
Query/avg_over_time(a_2000[1m]),_range_query_with_100_steps/Mimir-10                                                25.21m ±  1%   25.33m ±  4%        ~ (p=0.240 n=6)
Query/avg_over_time(a_2000[1m]),_range_query_with_1000_steps/Mimir-10                                               139.0m ±  3%   138.9m ±  4%        ~ (p=1.000 n=6)
Query/avg_over_time(nh_1[1m]),_instant_query/Mimir-10                                                               140.0µ ±  3%   136.9µ ±  1%        ~ (p=0.065 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_100_steps/Mimir-10                                                  411.3µ ±  5%   398.9µ ±  2%   -3.01% (p=0.041 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_1000_steps/Mimir-10                                                 1.748m ±  4%   1.669m ±  3%   -4.53% (p=0.009 n=6)
Query/avg_over_time(nh_100[1m]),_instant_query/Mimir-10                                                             650.8µ ± 10%   654.2µ ±  0%        ~ (p=0.132 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_100_steps/Mimir-10                                                21.45m ±  6%   21.27m ±  0%   -0.81% (p=0.002 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_1000_steps/Mimir-10                                               143.3m ±  3%   143.3m ±  1%        ~ (p=0.818 n=6)
Query/avg_over_time(nh_2000[1m]),_instant_query/Mimir-10                                                            8.182m ±  4%   8.318m ±  4%   +1.67% (p=0.041 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_100_steps/Mimir-10                                               405.2m ±  2%   404.7m ±  2%        ~ (p=0.485 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_1000_steps/Mimir-10                                               2.915 ±  3%    2.872 ±  2%   -1.49% (p=0.026 n=6)
Query/sum_over_time(a_1[1m]),_instant_query/Mimir-10                                                                131.2µ ± 10%   131.6µ ± 15%        ~ (p=0.485 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_100_steps/Mimir-10                                                   149.1µ ±  8%   151.3µ ±  3%        ~ (p=0.818 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_1000_steps/Mimir-10                                                  208.3µ ±  4%   209.4µ ±  4%        ~ (p=0.589 n=6)
Query/sum_over_time(a_100[1m]),_instant_query/Mimir-10                                                              383.6µ ±  4%   388.6µ ±  4%        ~ (p=0.132 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_100_steps/Mimir-10                                                 1.487m ±  3%   1.475m ±  4%   -0.78% (p=0.026 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_1000_steps/Mimir-10                                                6.733m ±  0%   6.706m ±  5%        ~ (p=0.065 n=6)
Query/sum_over_time(a_2000[1m]),_instant_query/Mimir-10                                                             4.153m ±  5%   4.205m ±  4%   +1.25% (p=0.041 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_100_steps/Mimir-10                                                23.45m ±  3%   23.39m ±  0%        ~ (p=0.485 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_1000_steps/Mimir-10                                               122.0m ±  4%   122.1m ±  1%        ~ (p=0.937 n=6)
Query/sum_over_time(nh_1[1m]),_instant_query/Mimir-10                                                               136.0µ ±  4%   134.6µ ±  1%        ~ (p=0.180 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_100_steps/Mimir-10                                                  294.0µ ±  5%   285.8µ ±  1%   -2.81% (p=0.015 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_1000_steps/Mimir-10                                                 850.6µ ±  1%   836.5µ ±  1%   -1.66% (p=0.002 n=6)
Query/sum_over_time(nh_100[1m]),_instant_query/Mimir-10                                                             572.9µ ±  0%   574.0µ ±  0%        ~ (p=0.180 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_100_steps/Mimir-10                                                13.14m ±  1%   13.01m ±  0%   -1.00% (p=0.002 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_1000_steps/Mimir-10                                               62.41m ±  3%   61.39m ±  3%   -1.64% (p=0.041 n=6)
Query/sum_over_time(nh_2000[1m]),_instant_query/Mimir-10                                                            6.644m ±  2%   6.684m ±  1%        ~ (p=0.132 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_100_steps/Mimir-10                                               247.5m ±  4%   246.7m ±  3%        ~ (p=0.394 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_1000_steps/Mimir-10                                               1.265 ±  1%    1.237 ±  1%   -2.21% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query/Mimir-10                                                         248.3µ ±  1%   251.9µ ±  2%   +1.47% (p=0.026 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps/Mimir-10                                            283.6µ ±  1%   287.6µ ±  3%   +1.42% (p=0.041 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps/Mimir-10                                           400.0µ ±  1%   405.9µ ±  3%        ~ (p=0.180 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query/Mimir-10                                                     858.1µ ±  4%   868.6µ ±  6%        ~ (p=0.065 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps/Mimir-10                                        2.884m ± 11%   2.925m ±  3%        ~ (p=0.093 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps/Mimir-10                                       13.17m ±  4%   13.23m ±  4%        ~ (p=0.485 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query/Mimir-10                                                   9.887m ±  4%   9.914m ±  2%        ~ (p=0.937 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps/Mimir-10                                      48.85m ±  1%   49.06m ±  2%        ~ (p=0.132 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps/Mimir-10                                     243.7m ±  5%   244.6m ±  4%        ~ (p=0.394 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_instant_query/Mimir-10                                                            154.9µ ±  2%   156.0µ ±  1%        ~ (p=0.093 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_100_steps/Mimir-10                                               231.4µ ±  3%   229.1µ ±  1%   -0.97% (p=0.026 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1000_steps/Mimir-10                                              593.1µ ±  2%   587.0µ ±  1%   -1.02% (p=0.041 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_instant_query/Mimir-10                                                          1.568m ±  3%   1.585m ±  0%        ~ (p=0.065 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_100_steps/Mimir-10                                             7.430m ±  3%   7.427m ±  1%        ~ (p=0.699 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_1000_steps/Mimir-10                                            36.56m ±  1%   36.59m ±  0%        ~ (p=0.937 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_instant_query/Mimir-10                                                         26.58m ±  3%   27.21m ±  3%   +2.39% (p=0.041 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_100_steps/Mimir-10                                            141.1m ±  4%   143.4m ±  1%        ~ (p=0.065 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_1000_steps/Mimir-10                                           702.5m ±  2%   702.5m ±  0%        ~ (p=0.937 n=6)
Query/sum_without_(l)(rate(a_1[1m])),_instant_query/Mimir-10                                                        137.3µ ± 13%   132.7µ ±  2%   -3.35% (p=0.026 n=6)
Query/sum_without_(l)(rate(a_1[1m])),_range_query_with_100_steps/Mimir-10                                           154.4µ ±  3%   154.3µ ±  5%        ~ (p=0.818 n=6)
Query/sum_without_(l)(rate(a_1[1m])),_range_query_with_1000_steps/Mimir-10                                          212.9µ ±  5%   210.3µ ±  1%        ~ (p=0.093 n=6)
Query/sum_without_(l)(rate(a_100[1m])),_instant_query/Mimir-10                                                      398.1µ ±  6%   392.6µ ±  2%        ~ (p=0.589 n=6)
Query/sum_without_(l)(rate(a_100[1m])),_range_query_with_100_steps/Mimir-10                                         1.453m ±  1%   1.453m ±  1%        ~ (p=0.937 n=6)
Query/sum_without_(l)(rate(a_100[1m])),_range_query_with_1000_steps/Mimir-10                                        6.488m ±  3%   6.386m ±  4%        ~ (p=0.180 n=6)
Query/sum_without_(l)(rate(a_2000[1m])),_instant_query/Mimir-10                                                     4.247m ±  2%   4.289m ±  7%        ~ (p=0.394 n=6)
Query/sum_without_(l)(rate(a_2000[1m])),_range_query_with_100_steps/Mimir-10                                        23.08m ±  2%   22.92m ±  1%        ~ (p=0.065 n=6)
Query/sum_without_(l)(rate(a_2000[1m])),_range_query_with_1000_steps/Mimir-10                                       116.8m ±  1%   116.3m ±  3%        ~ (p=0.394 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_instant_query/Mimir-10                       253.7µ ±  5%   255.3µ ±  4%        ~ (p=0.589 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_range_query_with_100_steps/Mimir-10          295.5µ ± 13%   290.2µ ±  2%   -1.81% (p=0.048 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_range_query_with_1000_steps/Mimir-10         420.8µ ±  3%   421.0µ ±  3%        ~ (p=0.818 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_instant_query/Mimir-10                   815.5µ ±  6%   820.0µ ±  6%        ~ (p=0.589 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_range_query_with_100_steps/Mimir-10      2.718m ±  6%   2.683m ±  1%   -1.28% (p=0.015 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_range_query_with_1000_steps/Mimir-10     12.19m ±  2%   12.11m ±  2%        ~ (p=0.093 n=6)
Query/sum_without_(l)(rate(a_2000[1m]))_/_sum_without_(l)(rate(b_2000[1m])),_instant_query/Mimir-10                 8.452m ±  1%   8.736m ±  1%   +3.37% (p=0.002 n=6)
…