@charleskorn 2024-08-18
Thinking about this more, I realised the solution isn't quite as simple as this, so I'm going to put this PR back to a draft until I can implement this properly:

It's not enough to check for duplicate series, what we need to do is check for conflicting samples for the same output series, and merge the results from input series that map to the same output series.

eg. imagine we're running the query `ceil({__name__=~"some_metric|some_other_metric"})`.

If the input data is:

```
some_metric        1 2 _ _
some_other_metric  _ _ 3 4
```

then we should not produce an error and should merge the result to a single series:

```
{} 1 2 3 4
```

However, the implementation here will cause the query to return an error. 

And, without this PR, we'll currently return this result, which is also invalid:

```
{} 1 2 _ _
{} _ _ 3 4
```

The only time we should return an error is if the input series map to the same output series and have a conflicting sample, for example:

```
some_metric       1 2 _ _
some_other_metric 1 _ _ _
```