@charleskorn 2024-08-30
Given the approval, I'm going to merge this, but happy to address any post-merge feedback in a follow-up PR.